Fixed PolygonMesher file name
