+++ src/foundry/client/pixi/core/shapes/index.d.ts
@@ -1,6 +1,6 @@
 import "./limited-angle-polygon";
 import "./polygon-helpers";
-import "./PolygonMesher";
+import "./polygon-mesher";
 import "./precise-text";
 import "./ray";
 import "./source-polygon";
